+++ client/src/entities/interview_question/index.ts
@@ -3,4 +3,5 @@ export * as interviewMutation from './mutations';
 export * as interviewQueries from './queries';
 export * as interviewType from './type';
 export { default as useInterViewStore } from './useInterViewStore';
+export { default as useRecordInfoStore } from './useRecordInfoStore';
 export { default as queryKeys } from './queryKeys';
