+++ client/src/features/interview-question/expected-question/ExpectedQuestionSelector.tsx
@@ -1,24 +1,40 @@
+import { styled } from 'styled-components';
+
+import { useCallback } from 'react';
+
+import {
+  interviewQueries,
+  useInterViewStore,
+} from '@/src/entities/interview_question';
+import { useDeleteInterViewQ } from '@/src/entities/interview_question/mutations';
 import { useCheckList } from '@/src/shared/hooks';
 import { Container } from '@/src/shared/components';
-import { styled } from 'styled-components';
+
+import { useSelectorLogic } from '../hooks';
+
 import QuestionSelector from './QuestionSelector';
 import ExpectedQuestionList from './ExpectedQuestionList';
 import ActionBtns from './ActionBtns';
 
-import useEUserQuestionManager from '../hooks/useEUserQuestionManager';
-import useExpectedQuestionManager from '../hooks/useExpectedQuestionManager';
-
 function ExpectedQuestionSelector() {
-  const { interViewQuestionList, dispatchRemoveQuestions } = useExpectedQuestionManager();
+  const { selectedType } = useInterViewStore();
+  const { data: interViewQuestionList = [] } = interviewQueries.useFetchQuestions(selectedType);
+  const { mutate: deleteQuestions } = useDeleteInterViewQ();
 
   const {
     checkList, handleAllCheck, handleCheckList, isAllCheck,
   } = useCheckList(interViewQuestionList);
 
-  const
-    { selectedType, dispatchSelectedType, dispatchSelectedQuestions } = useEUserQuestionManager({
-      typeCheckCallback: isAllCheck ? handleAllCheck : () => {},
-    });
+  const { dispatchSelectedType, dispatchSelectedQuestions } = useSelectorLogic({
+    typeCheckCallback: isAllCheck ? handleAllCheck : () => {},
+  });
+
+  const removeQuestions = useCallback((targetIds: number[]) => {
+    const formmatData = {
+      ids: targetIds,
+    };
+    deleteQuestions(formmatData);
+  }, []);
 
   return (
     <Container.ArticleCard
@@ -38,7 +54,7 @@ function ExpectedQuestionSelector() {
         checkList={checkList}
         onAdd={dispatchSelectedQuestions}
         isAllChecked={isAllCheck}
-        onRemove={dispatchRemoveQuestions}
+        onRemove={removeQuestions}
         onToggleAll={handleAllCheck}
         questions={interViewQuestionList}
       />
